Add required library so files for x265
This library file is required during linking by libavcodec when building ffmpeg. The latest version of x265 comes with a higher version of this file and the linker will not accept it. If this file is missing then the ffmpeg build script will download and extract this file as needed.